Add DestinationRepo class and tests
