--- src/Destinations.js
@@ -1,8 +1,0 @@
-class Destination {
-    constructor(data){
-        this.destinations = data
-    }
-    //retreive specific destinations method by id
-   
-}
-export default Destination
