+++ src/DestinationsRepo.js
@@ -3,7 +3,8 @@ class DestinationsRepo {
         this.allDestinationData = allDestinationData
     }
     getDestinationById(id){
-        let destFilter = this.destinations.filter(dest => id === dest.id)
+        let destFilter = this.allDestinationData.filter(dest => id === dest.id)
         return destFilter[0];
     }
 }
+export default DestinationsRepo;
