+++ src/Trips.js
@@ -12,14 +12,7 @@ class Trip {
         this.destination = destination
         this.cost;
     }
-    calculateTripCost() {
-        let lodging = this.destination.estimatedLodgingPerDay * this.duration;
-        let flight = this.destination.estimatedFlightCostPerPerson * this.travelers;
-        let tripCost = lodging + flight;
-        let agentFee = tripCost * 1.10;
-            this.cost = agentFee
-        return agentFee;
-        }
+    
     //method to return specific user id push to relevant array 
     //new Date() W3 /MDN 
     //method calculate the cost of the trip
